+++ src/main/java/insane96mcp/iguanatweaksreborn/IguanaTweaksReborn.java
@@ -5,6 +5,7 @@ import insane96mcp.iguanatweaksreborn.command.ITRCommand;
 import insane96mcp.iguanatweaksreborn.data.criterion.ITRTriggers;
 import insane96mcp.iguanatweaksreborn.data.generator.ITRBlockTagsProvider;
 import insane96mcp.iguanatweaksreborn.data.generator.ITRDamageTypeTagsProvider;
+import insane96mcp.iguanatweaksreborn.data.generator.ITREntityTypeTagsProvider;
 import insane96mcp.iguanatweaksreborn.data.generator.ITRItemTagsProvider;
 import insane96mcp.iguanatweaksreborn.data.generator.client.ITRBlockModelsProvider;
 import insane96mcp.iguanatweaksreborn.data.generator.client.ITRBlockStatesProvider;
@@ -135,6 +136,7 @@ public class IguanaTweaksReborn
         generator.addProvider(event.includeServer(), blockTags);
         generator.addProvider(event.includeServer(), new ITRItemTagsProvider(generator.getPackOutput(), lookupProvider, blockTags.contentsGetter(), IguanaTweaksReborn.MOD_ID, existingFileHelper));
         generator.addProvider(event.includeServer(), new ITRDamageTypeTagsProvider(generator.getPackOutput(), lookupProvider, IguanaTweaksReborn.MOD_ID, existingFileHelper));
+        generator.addProvider(event.includeServer(), new ITREntityTypeTagsProvider(generator.getPackOutput(), lookupProvider, IguanaTweaksReborn.MOD_ID, existingFileHelper));
         generator.addProvider(event.includeClient(), new ITRBlockStatesProvider(generator.getPackOutput(), IguanaTweaksReborn.MOD_ID, existingFileHelper));
         generator.addProvider(event.includeClient(), new ITRBlockModelsProvider(generator.getPackOutput(), IguanaTweaksReborn.MOD_ID, existingFileHelper));
         generator.addProvider(event.includeClient(), new ITRItemModelsProvider(generator.getPackOutput(), IguanaTweaksReborn.MOD_ID, existingFileHelper));